Remove CLI enum of devkit
