+++ horusec-cli/internal/enums/outputtype/output_type.go
@@ -12,15 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cli
+package outputtype
 
-import (
-	"github.com/stretchr/testify/assert"
-	"testing"
+type OutputType string
+
+const (
+	Text      OutputType = "text"
+	JSON      OutputType = "json"
+	SonarQube OutputType = "sonarqube"
 )
 
-func TestToString(t *testing.T) {
-	t.Run("Should success parse to string", func(t *testing.T) {
-		assert.Equal(t, "json", JSON.ToString())
-	})
+func (o OutputType) ToString() string {
+	return string(o)
 }
