+++ horusec-cli/internal/usecases/cli/cli_test.go
@@ -15,9 +15,9 @@
 package cli
 
 import (
+	"github.com/ZupIT/horusec/horusec-cli/internal/enums/outputtype"
 	"testing"
 
-	"github.com/ZupIT/horusec/development-kit/pkg/enums/cli"
 	cliConfig "github.com/ZupIT/horusec/horusec-cli/config"
 	"github.com/ZupIT/horusec/horusec-cli/internal/entities/workdir"
 	"github.com/stretchr/testify/assert"
@@ -72,7 +72,7 @@ func TestValidateConfigs(t *testing.T) {
 		config := &cliConfig.Config{}
 		config.SetWorkDir(&workdir.WorkDir{})
 		config.NewConfigsFromEnvironments()
-		config.SetPrintOutputType(cli.JSON.ToString())
+		config.SetPrintOutputType(outputtype.JSON.ToString())
 		config.SetJSONOutputFilePath("")
 
 		err := useCases.ValidateConfigs(config)
@@ -85,7 +85,7 @@ func TestValidateConfigs(t *testing.T) {
 		config := &cliConfig.Config{}
 		config.SetWorkDir(&workdir.WorkDir{})
 		config.NewConfigsFromEnvironments()
-		config.SetPrintOutputType(cli.JSON.ToString())
+		config.SetPrintOutputType(outputtype.JSON.ToString())
 		config.SetJSONOutputFilePath("test.test")
 
 		err := useCases.ValidateConfigs(config)
